rotation tool: tune suggested center position
During the preparation of the rotation tool dialog,
the rotation center is placed exactly in the middle of
the box. In the case of a +/-90 degrees rotation, the
result is more blurry unless the center is in the corner
of a pixle or in the middle of a pixel. This places the
suggested rotation center position always on a pixel
corner. See issue description on:
https://gitlab.gnome.org/GNOME/gimp/issues/3030