build: fix cross-compiled meson builds.
A bug came with meson 0.54.1 (looking at reports, I understand it to be
a mix of a meson bug and a patch by Debian making it a different bug).
See:
- https://github.com/mesonbuild/meson/issues/6115#issuecomment-593312709
- https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=959708

In any case, our Windows cross-builds end up with (with no libdir set on
our own): libdir=${prefix}/lib/x86_64-linux-gnu
This is obviously wrong as confirmed by meson report comments.