Add missing includes
We were exposing symbols in some headers, where we neglected to include
the symbol's header itself. This was okay for gcc because the header was
included _somewhere_, but not at the right place. The GIR parser tries
to resolve all symbols in a given header file by recursively looking in
all included files, so if we don't import the correct headers, the GIR
scanner will give errors.