plug-ins: Use pygobject's auto GValue coversion
In the previous commit, we introuced API that lets us use an array of
`GValue`s rather than a `GimpValueArray` to make the bindings feel more
native. In this commit, we go a step further and use the ability of
pygobject to do automatic GValue conversion to make the API even
simpler (note that this possibly might not work in each language binding
though).