gimpprocedure: Use less strict typechecking
When executing `GimpProcedure.run()`, we perform some type checking
(and sometimes even do more validation) to make sure that plug-ins are
behaving correctly.

The problem with the current implementation was that a plug-in that
accepted for example a single `GimpDrawable` as argument, then trying to
call this with a value of `GIMP_TYPE_CHANNEL` didn't pass the
validation, even though the latter is just a subclass of the former.

For that reason, do a proper typecheck using `g_type_is_a()`.