app: make "gegl:matting-levin" default engine of Foreground Select tool.
Even though this engine is optional, we already have the code to detect
its absence at runtime, and to fallback to "gegl:matting-global". So it
won't be a problem even then.

When the operation is present though, it definitely makes a lot more
sense than matting global as default, because it performs a lot better
in most cases (as far as I could see as well as others).

(cherry picked from commit b52cdf4c90800452aa9717d57398ba72a5819d2a)