build: add an additional distcheck step to .gitlab-ci.yml
Runs in the gimp stage in parallel to the make && make check build.
May catch dist errors.

Distcheck is allowed to fail. The idea is that being able to build GIMP
is more important and should get fast feedback, while make distcheck is
usually only used when a release is prepared.

Reviewer note: not sure about the "allowed to fail" part because we also
want to fix dist bugs as soon as we can (it's easier to find a bug
source when we get notified as soon as a guilty commit breaks the CI).
But anyway this is a step forward from current CI. So let's merge this
for now.