plug-ins: fix missing "plug-in-colormap-swap" procedure.
Existing code was returning only 1 out of 2 expected procedure names.

See !241 for the originally proposed patch by Rafał Mikrut (@qarmin).
Unfortunately the proposed patch also had another bug and was always
returning an empty list instead, and the contributor is unresponsive. So
let's just redo the patch and be done with it.

Also using g_list_prepend() instead of g_list_append() as it's usually
more efficient (to be fair here, for a size-2 list, it doesn't matter
much but it's good practice anyway) and order doesn't matter.

Thanks Rafał Mikrut for noticing the original bug!