Issue #3481: Cannot paste images into GIMP copied from Telegram.
Adding patch by Massimo Valentini (@Massimo) for what is seemingly a bug
in GTK+2.24 in Windows.

The bug does not seem to happen on GIMP 3 (according to contributor
tests), so the bug/patch is likely not relevant anymore on GTK+3. So
let's just add this patch in our official GIMP build and be done with
it.